test(search): separate composition coverage and verify the global shortcut
