+++ ods/extensions/services/dashboard/src/components/PixelCommandSearch.test.jsx
@@ -30,19 +30,3 @@ test('handles empty results without issuing an action', () => {
   expect(create).not.toHaveBeenCalled()
   expect(screen.getByText('No matching conversations or actions.')).toBeVisible()
 })
-
-test('leaves IME candidate keys to text composition before activating a result', () => {
-  const create = vi.fn()
-  render(<MemoryRouter><PixelCommandSearch onInsert={() => {}} onNewTask={create}/></MemoryRouter>)
-  fireEvent(window, new Event(OPEN_PIXEL_SEARCH))
-  const input = screen.getByLabelText('Search conversations and actions')
-  fireEvent.compositionStart(input)
-  expect(fireEvent.keyDown(input, {key:'ArrowDown',isComposing:true})).toBe(true)
-  expect(fireEvent.keyDown(input, {key:'Enter',isComposing:true})).toBe(true)
-  expect(create).not.toHaveBeenCalled()
-  expect(screen.getByRole('dialog')).toBeVisible()
-  fireEvent.compositionEnd(input)
-  fireEvent.keyDown(input, {key:'Enter'})
-  expect(create).toHaveBeenCalledOnce()
-  expect(screen.queryByRole('dialog')).toBeNull()
-})
